python-fast-reply-service/requirements.txt
# Core dependencies
redis>=5.0.0
aiosmtplib>=3.0.0
imap-tools>=1.5.0
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# CrewAI for AI processing
crewai>=0.11.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-deepseek>=0.1.0

# Supabase for database
supabase>=2.0.0

# Google APIs (for Gmail Pub/Sub)
google-api-python-client>=2.108.0
google-auth>=2.23.0
google-cloud-pubsub>=2.18.0

# Microsoft Graph (for Outlook)
msal>=1.24.0

# Async utilities
asyncio-mqtt>=0.16.0
aiohttp>=3.9.0

# Monitoring
prometheus-client>=0.19.0
structlog>=23.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
